requirements.txt
configparser==4.0.2
Django==2.2.6
django-cors-headers==2.4.0
djangorestframework==3.10.3
mysqlclient==1.4.4